# Improve add mesh performance