… tests/plotting/image_cache/bool_scalars.png
… tests/test_colors.py
… tests/test_plotter.py