This PR improves the performance of [add_mesh](https://docs.pyvista.org/api/plotting/_autosummary/pyvista.Plotter.add_mesh.html) by about 50% on Linux and 60% Windows by cutting out various unnecessary calls.

#### Why?

Fundamentally, a few milliseconds per `add_mesh` isn't so bad. That is, unless you're calling it a thousand times.

Ideally you would merge all these individual datasets and add a single dataset to the plotter, but that means you can't individually pick each dataset via an actor.

Trust me, I don't want to do this, but as it stands, based on an internal project, reducing the bottleneck is the best I can do.


## Test Setup and Timings

Note that this is not the ideal approach to add identical meshes. Use [glyph()](https://docs.pyvista.org/api/core/_autosummary/pyvista.DataSet.glyph.html) for a more efficient implementation.

```py
import numpy as np
import pyvista as pv

meshes = []
for i in range(1000):
    mesh = pv.Sphere(center=np.random.random(3)*10)
    mesh['data'] = np.random.random(mesh.n_points)
    meshes.append(mesh)


def add_meshes():
    pl = pv.Plotter()
    for mesh in meshes:
        pl.add_mesh(mesh)
    pl.close()

```

### Linux: `main`
```py
>>> timeit add_meshes()
2.91 s ± 28.6 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

```

### Linux: This Branch

```py
>>> timeit add_meshes()
1.34 s ± 3.76 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

### Windows: `main`
```py
>>> timeit add_meshes()
4.6 s ± 180 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

### Windows: This Branch
```py
>>> timeit add_meshes()
1.8 s ± 20.3 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```