@adeak 2022-07-13
I'm planning to look at this and some other PRs (e.g. https://github.com/pyvista/pyvista/pull/2980, https://github.com/pyvista/pyvista/pull/2964), but I won't have time for that until tomorrow or maybe even Friday. I see there's a varying degree of rush here, I should probably start with the 0.35.2-milestoned PR (assuming they're still open when I get there).