@akaszynski 2022-07-17
Breaking this up into:
- [ ] Cache attempted imports
- [ ] Refactor mapper
- [ ] Fix camera reset

Leaving this up for now.